requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-firestore==2.13.1
pydantic==2.5.0
jinja2==3.1.2
python-multipart==0.0.6